chore(tests): move app bundle setup helpers to tests/common.ts
